feat: #1 Room 모델 및 서비스, 메모리저장 구현

+++ sigkill-server/src/main/java/com/gulab/sigkillserver/domain/room/dto/response/RoomListResponse.java
@@ -9,7 +9,7 @@ public record RoomListResponse(
         List<RoomResponse> rooms,
         Integer page,
         Integer size,
-        Long totalElements,
+        Integer totalElements,
         Integer totalPages,
         Boolean hasNext
 ) {
